Adding source code provided by the course
This commit contains the source code needed for the assignments in this course. This material is needed to finish each assignment, and is not authored by me.

ALL files in this commit are covered by the respective license terms set by the creators of these files (Andrew Ng, Stanford, Coursera etc.).